# api/openapi.yaml
openapi: 3.1.0
info:
  title: Spoiler-Safe TV Q&A API
  version: 0.1.0
  description: |
    Single source of truth. v0 contains only health endpoints; real endpoints
    are added in subsequent backlog items.
servers:
  - url: http://localhost:8081
    description: user-progress (local)
  - url: http://localhost:8082
    description: catalog (local)
  - url: http://localhost:8083
    description: chat (local)
  - url: http://localhost:8084
    description: genai (local)
paths:
...
  /user-progress/likes:
    get:
      tags: [likes]
      operationId: listLikes
      summary: List the series the current user has liked ("My Shows"), newest first
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Series IDs the current user has liked, most recently liked first
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
                  format: uuid
        '401':
          description: Missing or invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /user-progress/likes/{seriesId}:
    put:
      tags: [likes]
      operationId: likeSeries
      summary: Like a series (idempotent — liking an already-liked series is a no-op)
      security:
        - bearerAuth: []
      parameters:
        - name: seriesId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: The series is liked
        '401':
          description: Missing or invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      tags: [likes]
      operationId: unlikeSeries
      summary: Remove a like (idempotent — unliking a series that isn't liked is a no-op)
      security:
        - bearerAuth: []
      parameters:
        - name: seriesId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: The series is not liked
        '401':
          description: Missing or invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT bearer token. Obtain from /user-progress/auth/login
...
  schemas:
...
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          example: VALIDATION_FAILED
        message:
          type: string
        details:
          type: object
          additionalProperties: true
...